+++ src/App.js
@@ -1,11 +1,11 @@
 import React from 'react'
 import GlobalStyles from './GlobalStyles'
 import Header from './Header'
-import Home from './Home'
-import TaskList from './TaskList'
-import IndividualNotes from './IndividualNotes'
-import CreateShipment from './CreateShipment'
-import MyShipments from './MyShipments'
+import Home from './HomeComponent/Home'
+import TaskList from './GuideComponent/TaskList'
+import IndividualNotes from './NoteComponent/IndividualNotes'
+import CreateShipment from './CreateComponent/CreateShipment'
+import MyShipments from './MyShipmentComponent/MyShipments'
 import Footer from './Footer'
 import { Switch, Route } from 'react-router-dom'
