+++ src/GuideComponent/GetNextTaskButton.stories.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { action } from '@storybook/addon-actions'
 import GetNextTaskButton from './GetNextTaskButton'
-import withMobileWrapper from '../.storybook/preview'
+import withMobileWrapper from '../../.storybook/preview'
 
 export default {
   component: 'GetNextTaskButton',
