+++ src/GuideComponent/ProgressBar.js
