+++ src/GuideComponent/TaskList.js
@@ -2,25 +2,25 @@ import React, { useState, useEffect } from 'react'
 import styled from 'styled-components/macro'
 import GetNextTaskButton from './GetNextTaskButton'
 import { Link, useRouteMatch } from 'react-router-dom'
-import info from './images/info.jpg'
-import product from './images/product.jpg'
-import numbers from './images/numbers.jpg'
-import restriction from './images/restriction.jpg'
-import sample from './images/sample.jpg'
-import dollar from './images/dollar.jpg'
-import calculation from './images/calculation.jpg'
-import contract from './images/contract.jpg'
-import eori from './images/eori.jpg'
-import container from './images/container.jpg'
-import money from './images/money.jpg'
-import data from './images/data.jpg'
-import pallets from './images/pallets.jpg'
-import tax from './images/tax.jpg'
-import deal from './images/deal.jpg'
-import congratulations from './images/congratulations.jpg'
+import info from '../images/info.jpg'
+import product from '../images/product.jpg'
+import numbers from '../images/numbers.jpg'
+import restriction from '../images/restriction.jpg'
+import sample from '../images/sample.jpg'
+import dollar from '../images/dollar.jpg'
+import calculation from '../images/calculation.jpg'
+import contract from '../images/contract.jpg'
+import eori from '../images/eori.jpg'
+import container from '../images/container.jpg'
+import money from '../images/money.jpg'
+import data from '../images/data.jpg'
+import pallets from '../images/pallets.jpg'
+import tax from '../images/tax.jpg'
+import deal from '../images/deal.jpg'
+import congratulations from '../images/congratulations.jpg'
 import ProgressBar from './ProgressBar'
 import PropTypes from 'prop-types'
-import { db } from './firebase'
+import { db } from '../firebase'
 
 ProgressBar.propTypes = {
   percentage: PropTypes.number,
