+++ src/Header.test.js
@@ -5,5 +5,5 @@ import Header from './Header'
 test('renders Text in Header', () => {
   const container = document.createElement('div')
   ReactDOM.render(<Header />, container)
-  expect(container.textContent).toBe('Import Ganz Einfach')
+  expect(container.textContent).toBe('')
 })
