+++ src/HomeComponent/Home.stories.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import Home from './Home'
-import withMobileWrapper from '../.storybook/preview'
+import withMobileWrapper from '../../.storybook/preview'
 
 export default {
   component: 'Home',
