+++ src/NoteComponent/NoteForm.stories.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import NoteForm from './NoteForm'
-import withMobileWrapper from '../.storybook/preview'
+import withMobileWrapper from '../../.storybook/preview'
 
 export default {
   component: 'NoteForm',
